requirements.txt
kivy==2.3.1
pydantic==2.10.6
openai==1.65.4
python-dotenv==0.9.9
kivymd==1.2.0